docs(examples): Service examples refactor, user-avatar examples added
