+++ src/playground/menu/menu-service.component.scss
@@ -1,0 +1,8 @@
+[nbButton] {
+  margin-right: 1rem;
+  margin-bottom: 1rem;
+}
+h3 {
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+}
