+++ src/playground/user/user-avatar-settings.component.html
@@ -1,0 +1,50 @@
+<nb-card>
+  <nb-card-body>
+    <p>User avatar image set as link</p>
+    <nb-user size="large"
+             name="Jonh Doe"
+             title="Engineer"
+             picture="https://via.placeholder.com/50/4479e7/ffffff?Text=IMG">
+    </nb-user>
+
+  </nb-card-body>
+</nb-card>
+
+<nb-card>
+  <nb-card-body>
+
+    <p>User avatar image set as BASE64 string</p>
+    <nb-user size="large"
+             name="Jonh Doe"
+             title="Engineer"
+             [picture]="base64image">
+    </nb-user>
+
+  </nb-card-body>
+</nb-card>
+
+<nb-card>
+  <nb-card-body>
+
+    <p>User avatar image not set and showInitials disabled</p>
+    <nb-user size="large"
+             name="Jonh Doe"
+             title="Engineer"
+             showInitials=false>
+    </nb-user>
+
+  </nb-card-body>
+</nb-card>
+
+<nb-card>
+  <nb-card-body>
+
+    <p>User avatar with custom background color</p>
+    <nb-user size="large"
+             name="Jonh Doe"
+             title="Engineer"
+             color="#cccccc">
+    </nb-user>
+
+  </nb-card-body>
+</nb-card>
