+++ src/playground/user/user-hide-captions.component.html
@@ -1,32 +1,11 @@
 <nb-card>
   <nb-card-body>
 
-    <nb-user size="small"
-             name="Jonh Doe"
-             title="Engineer">
-    </nb-user>
-
-  </nb-card-body>
-</nb-card>
-
-
-<nb-card>
-  <nb-card-body>
-
-    <nb-user size="middle"
-             name="Jonh Doe"
-             title="Engineer">
-    </nb-user>
-
-  </nb-card-body>
-</nb-card>
-
-<nb-card>
-  <nb-card-body>
-
+    <p>Only picture</p>
     <nb-user size="large"
              name="Jonh Doe"
-             title="Engineer">
+             title="Engineer"
+             onlyPicture>
     </nb-user>
 
   </nb-card-body>
@@ -35,9 +14,24 @@
 <nb-card>
   <nb-card-body>
 
-    <nb-user size="xlarge"
+    <p>Name hidden</p>
+    <nb-user size="large"
              name="Jonh Doe"
-             title="Engineer">
+             title="Engineer"
+             showName=false>
+    </nb-user>
+
+  </nb-card-body>
+</nb-card>
+
+<nb-card>
+  <nb-card-body>
+
+    <p>Title hidden</p>
+    <nb-user size="large"
+             name="Jonh Doe"
+             title="Engineer"
+             showTitle=false>
     </nb-user>
 
   </nb-card-body>
