+++ src/playground/user/user-hide-captions.component.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * Copyright Akveo. All Rights Reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'nb-user-hide-captions',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  templateUrl: './user-hide-captions.component.html',
+})
+export class NbUserHideCaptionsComponent {
+}
